+++ changelog.d/42.changed.md
@@ -1,0 +1,1 @@
+Split the oversized `aces_sdl.validator` module (4,139 lines) into a package of per-validation-seam mixin modules (`_core`, runtime families, relationships, content/objectives, workflows, sections), each under the ADR-015 600-line cap, behind an API-stable `SemanticValidator` re-export. Pure refactor: no validation behavior, diagnostics, pass ordering, or public-API change. The `validator.py` entry is removed from the oversized-source allowlist.
