+++ implementations/python/packages/aces_sdl/validator/_support.py
@@ -1,0 +1,30 @@
+"""Shared module-level helpers for the SemanticValidator package."""
+
+from collections import defaultdict, deque
+
+# Common ref-path prefix used by qualified runtime/service refs (e.g.
+# ``nodes.vm.services.http``, ``nodes.vm.runtime.applications.webapp``).
+_NODES_PREFIX = "nodes."
+
+
+def _topological_sort(graph: dict[str, list[str]]) -> list[str] | None:
+    """Return topological order or None if a cycle exists."""
+    in_degree: dict[str, int] = defaultdict(int)
+    for node in graph:
+        in_degree.setdefault(node, 0)
+    for deps in graph.values():
+        for dep in deps:
+            in_degree[dep] += 1
+
+    queue = deque(n for n, d in in_degree.items() if d == 0)
+    order: list[str] = []
+
+    while queue:
+        node = queue.popleft()
+        order.append(node)
+        for dep in graph.get(node, []):
+            in_degree[dep] -= 1
+            if in_degree[dep] == 0:
+                queue.append(dep)
+
+    return order if len(order) == len(in_degree) else None
