+++ implementations/python/pyproject.toml
@@ -131,4 +131,4 @@ ignore = [
 "packages/aces_runtime/control_plane_api.py" = ["B008"]  # fastapi Depends() pattern
 "packages/aces_runtime/control_plane.py" = ["S112"]      # intentional exception suppression
 "packages/aces_sdl/module_registry.py" = ["S310", "S202"]  # explicit OCI URL fetch and tarball extract
-"packages/aces_sdl/validator.py" = ["F821"]                # string-quoted forward references in closures
+"packages/aces_sdl/validator/_sections.py" = ["F821"]      # string-quoted forward references in closures
