+++ specs/formal/objectives/window-consistency.md
@@ -69,7 +69,7 @@ shape before compiler/planner semantics run. Each resolved reference carries:
 ## Implementation Mapping
 
 - shared semantic source of truth: `implementations/python/packages/aces_sdl/semantics/objectives.py`
-- validator checks: `implementations/python/packages/aces_sdl/validator.py`
+- validator checks: `implementations/python/packages/aces_sdl/validator/`
 - compiled runtime references and refresh derivation:
   - `implementations/python/packages/aces_processor/compiler.py`
   - `implementations/python/packages/aces_processor/models.py`
