+++ specs/formal/participant-semantics/README.md
@@ -716,7 +716,7 @@ Current implementation artifacts for the `SEM-208` slice:
 - `implementations/python/packages/aces_sdl/participant_behavior.py` defines
   typed action contracts and observation boundaries;
 - `implementations/python/packages/aces_sdl/semantics/participant_behavior.py`
-  and `implementations/python/packages/aces_sdl/validator.py` fail closed on
+  and `implementations/python/packages/aces_sdl/validator/` fail closed on
   unbound action-contract and observation-boundary references;
 - `implementations/python/packages/aces_processor/compiler.py` maps authored
   participants to compiled participant action, observation, and behavior
@@ -759,7 +759,7 @@ Current implementation artifacts for the `SEM-209` slice:
   interaction classes, target references, related actions, and shared-state
   references on action contracts;
 - `implementations/python/packages/aces_sdl/semantics/participant_behavior.py`
-  and `implementations/python/packages/aces_sdl/validator.py` fail closed on
+  and `implementations/python/packages/aces_sdl/validator/` fail closed on
   unbound related actions, interaction targets, and shared-state references;
 - `implementations/python/packages/aces_processor/compiler.py` carries declared
   interaction classes and shared-state references into compiled participant
@@ -830,7 +830,7 @@ Current implementation artifacts for the `SEM-210` slice:
   observation-boundary hidden, observable, and evidence-only reference
   separation;
 - `implementations/python/packages/aces_sdl/semantics/participant_behavior.py`
-  and `implementations/python/packages/aces_sdl/validator.py` continue to
+  and `implementations/python/packages/aces_sdl/validator/` continue to
   fail closed on unbound participant observation-boundary references, view-rule
   references, and view-transition evidence references;
 - `implementations/python/packages/aces_processor/compiler.py` carries hidden,
